+++ src/submissions/backend/excel/parsers/results_parsers/__init__.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 from pathlib import Path
 from backend.excel.parsers import DefaultKEYVALUEParser, DefaultTABLEParser
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Tuple, Generator, Any
 import logging
 if TYPE_CHECKING:
     from backend.db.models import ProcedureType
@@ -12,7 +12,7 @@ logger = logging.getLogger(f"submissions.{__name__}")
 class DefaultResultsInfoParser(DefaultKEYVALUEParser):
     pyd_name = "PydResults"
 
-    def __init__(self, filepath: Path | str, results_type: str, proceduretype: "ProcedureType" | None = None,
+    def __init__(self, filepath: Path | str, results_type: str, proceduretype: ProcedureType | None = None,
                   *args, **kwargs):
         if results_type:
             self.results_type = results_type
@@ -27,9 +27,17 @@ class DefaultResultsInfoParser(DefaultKEYVALUEParser):
                 start_row = 1
         else:
             start_row = kwargs.pop('start_row')
-        # start_row = proceduretype.allowed_result_methods[results_type]['info']['start_row']
         super().__init__(filepath=filepath, proceduretype=proceduretype, sheet=sheet, start_row=start_row, *args,
                          **kwargs)
+
+    @property
+    def parsed_info(self) -> Generator[Tuple[str, Any], None, None]:
+        for key, value in super().parsed_info:
+            try:
+                value = value['value']
+            except KeyError:
+                pass
+            yield key, value
 
 
 class DefaultResultsSampleParser(DefaultTABLEParser):